requirements.txt
Flask==3.0.3
Flask-Cors==4.1.2
azure-storage-blob==12.22.0
Werkzeug==3.0.4
gunicorn==23.0.0
python-dotenv==1.0.1